Fix more OS X things we need
Summary:
* compatibility for adding functions that are missing
* Added strndup from PHP
* uint -> unsigned int in zend_qsort
* hotprofiler sync from open source xhprof
* pthread_t isn't an unsigned long on OSX, explicit cast
* signals typo in process for macro

Test Plan:
make
make fast_tests

DiffCamp Revision: 142772
Reviewed By: hzhao
CC: hphp-diffs@lists, davidrecordon, hzhao
Revert Plan:
Ok

+++ src/runtime/ext/ext_thread.cpp
@@ -37,7 +37,7 @@ void f_hphp_thread_set_warmup_enabled() {
 }
 
 int64 f_hphp_get_thread_id() {
-  return Process::GetThreadId();
+  return  (unsigned long)Process::GetThreadId();
 }
 
 ///////////////////////////////////////////////////////////////////////////////
